Optimized IsPrime class
The is prime class has been optimized to run faster. The program now
checks if a number is divisible by two and then goes into a for loop
that starts at three and increments by two. The loop then stops when i
reaches the square root of the number we are checking.